Law 24: create Combinations (#49)
* data layer

* Wiring for upload - first working pass

* reasonable update to source list + buttons

* editable usage and combination display

* Edit single combination

* rename tabs

* fix tests

* first new test

* tests for combo list

* fix a bug in upload combos and get tests done

* pepdiff and LOC coverage

* First round (oy) on Example images

- add field for combination detail link
- make it so position is not the only game in town (not required)
- update bulk upload
- update image list (preliminary) to show top level data that includes the combo (y/n) and that if a position is not present, that's handled
- also needed admin for further verification

TODO - edit an example, add an example, details to include combos, have combo list include images.  Then to work on read only views.

* Update image edit/add/copy

now can include combo details or position, or both.

Also cleaned up display of position details to be more dynamic

* show actual combos on image list

if available, if not, no change.

* add pic to list

* fix bugs, add and clean up tests

* More test coverage

* more coverage

* pepdiff

* First pass at view only view of combos

* more tests

* coverage for positions w images

* pepdiff

* test another condition

* pepdiff

* pull print statements

* test coverage

* bug fix

fix the flow for when there are only combos in the image upload

And fix a residual bug, previously uncaught - that by redoing image list to be per image, we have translate exampleimages into the actual ids of images.  Easier to do this close to the source.

* tests for previous bug